dealflow_assistant/requirements.txt
beautifulsoup4==4.13.4
dotenv==0.9.9
fastapi==0.115.13
google-generativeai==0.8.5
lxml==5.4.0
python-dotenv==1.1.1
requests==2.32.4
selenium==4.33.0
streamlit==1.45.1
tqdm==4.67.1
uvicorn==0.34.3